+++ app/Http/Controllers/materialPublicoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\usuarios;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use App\Models\MaterialPublico;
@@ -15,9 +16,9 @@ class materialPublicoController extends Controller
 {
     public function index(Request $request)
     {
-        if($request->idUsuario){
+        if ($request->idUsuario) {
             $peticionSQL = $this->getMaterialPublicoForUsuario($request);
-        }else{
+        } else {
             $peticionSQL = $this->getMaterialPublico($request);
         }
 
@@ -26,9 +27,9 @@ class materialPublicoController extends Controller
 
         foreach ($peticionSQL as $p) {
             $peticionSQLFiltrada = $this->getArchivosMaterialPublico($p);
-
-            $p->imgEncabezado = base64_encode(Storage::disk('ftp')->get($p->imgEncabezado));
-
+            if (!App::environment(['testing'])) {
+                $p->imgEncabezado = base64_encode(Storage::disk('ftp')->get($p->imgEncabezado));
+            }
             $arrayArchivos = array();
 
 
@@ -58,17 +59,17 @@ class materialPublicoController extends Controller
     }
 
     public function store(Request $request)
-    {   
+    {
         $request->validate([
-            'titulo' => 'string|required', 
-            'mensaje' => 'string', 
+            'titulo' => 'string|required',
+            'mensaje' => 'string',
             'idUsuario' => 'required'
         ]);
         $usuario = usuarios::findOrFail($request->idUsuario);
-        if($usuario->ou != "Profesor"){
+        if ($usuario->ou != "Profesor") {
             return response()->json(['status' => 'Unauthorized'], 401);
         }
-        
+
         $idDatos = $this->agregarMaterialPublico($request);
 
 
@@ -83,7 +84,7 @@ class materialPublicoController extends Controller
         return response()->json(['status' => 'Success'], 200);
     }
 
-    public function destroy($id,Request $request)
+    public function destroy($id, Request $request)
     {
 
         $materialPublico = MaterialPublico::findOrFail($id);
@@ -135,7 +136,7 @@ class materialPublicoController extends Controller
         return $peticionSQLFiltrada;
     }
 
-  
+
     public function agregarMaterialPublico(Request $request)
     {
         $materialPublico = new MaterialPublico;
@@ -148,11 +149,13 @@ class materialPublicoController extends Controller
         return $materialPublico;
     }
 
-    
+
     public function subirArchivoMaterialPublico(Request $request, int $i, $idDatos)
     {
         $nombreArchivo = random_int(0, 1000000) . "_" . $request->nombresArchivo[$i];
-        Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        if (!App::environment(['testing'])) {
+            Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        }
         $archivosForo = new archivos_material_publico;
         $archivosForo->idMaterialPublico = $idDatos->id;
         $archivosForo->nombreArchivo = $nombreArchivo;
@@ -161,7 +164,9 @@ class materialPublicoController extends Controller
 
     public function deleteArchivosMaterialPublico($p)
     {
-        Storage::disk('ftp')->delete($p->nombreArchivo);
+        if (!App::environment(['testing'])) {
+            Storage::disk('ftp')->delete($p->nombreArchivo);
+        }
         $p->delete();
     }
 }
